fix(sonar): harden scenario path and clear backend service issues
Scenario files are checked against a file-name whitelist and a
normalised path under scenarios/ at the point of reading. The dump
import now runs as one JDBC batch and one sequence resync statement.

+++ src/main/java/dev/sylvain/planning/service/solve/LargeProblemConstruction.java
@@ -61,7 +61,7 @@ final class LargeProblemConstruction {
         if (problem == null || problem.getPostes() == null || problem.getAnimateurs() == null) {
             return false;
         }
-        return (long) problem.getPostes().size() * (problem.getAnimateurs().size() + 1L) >= THRESHOLD;
+        return problem.getPostes().size() * (problem.getAnimateurs().size() + 1L) >= THRESHOLD;
     }
 
     /** Replaces the first construction heuristic phase of {@code solverConfig} with the sampled one. */
